move test content into its own file
